Add pipeline_communicate() generator
This generator will be used as the "send" operation during execute()
step when the connection is in pipeline mode. It can consume results or
send queries depending on socket read/write ready state.

Added test is taken from PostgreSQL sources. It demonstrates the use of
pipeline mode where query-send and results-fetch steps are interleaved
without any sync point emitted. (In this test, this works because the
output buffer gets full.) The test writes data to logs. These will be
displayed upon failure by pytest but can be seen otherwise with
--log-file=<path> command-line argument.

Typically, when running:

    pytest tests/test_pipeline.py::test_pipeline_communicate --log-file=log

the 'log' file would contain (prefixes stripped):

    sent begin: BEGIN TRANSACTION
    sent drop table: DROP TABLE IF EXISTS pq_pipeline_demo
    sent create table: CREATE UNLOGGED TABLE pq_pipeline_demo( id serial primary key, itemno integer, int8filler int8)
    sent prepare: INSERT INTO pq_pipeline_demo(itemno, int8filler) VALUES ($1, $2)
    sent row 9999
    sent row 9998
    sent row 9997
    sent row 9996
    sent row 9995
    sent row 9994
    sent row 9993
    sent row 9992
    sent row 9991
    ...
    sent row 9690
    sent row 9689
    sent row 9688
    got COMMAND_OK results for 'begin' command
    got COMMAND_OK results for 'drop table' command
    got COMMAND_OK results for 'create table' command
    got COMMAND_OK results for 'prepare' command
    got COMMAND_OK results for 'insert 9999' command
    got COMMAND_OK results for 'insert 9998' command
    got COMMAND_OK results for 'insert 9997' command
    got COMMAND_OK results for 'insert 9996' command
    got COMMAND_OK results for 'insert 9995' command
    ...
    got COMMAND_OK results for 'insert 9691' command
    got COMMAND_OK results for 'insert 9690' command
    sent row 9687
    sent row 9686
    sent row 9685
    sent row 9684
    sent row 9683
    ...
    sent row 2
    sent row 1
    sent row 0
    sent COMMIT
    pipeline sync sent
    got COMMAND_OK results for 'insert 236' command
    got COMMAND_OK results for 'insert 235' command
    got COMMAND_OK results for 'insert 234' command
    got COMMAND_OK results for 'insert 233' command
    ...
    got COMMAND_OK results for 'insert 0' command
    got COMMAND_OK results for 'commit' command
    got PIPELINE_SYNC results for 'pipeline sync' command

We can see that commands are sent, until the output buffer is full (the
connection is then Read-ready only), then results are fetched, until
more commands can be sent, and the cycle repeat.